remove SAAS_PROTOTYPE toggle and delete prototype app
Move CacheStore and fake_data utilities into the styleguide app,
which is the only remaining consumer. Remove the prototype app
entirely including its flag-gated views, URLs, templates, and
static files.
