--- corehq/apps/prototype/models/__init__.py
@@ -1,4 +1,0 @@
-"""
-Leaving this as a module so that prototypes can organize their models here.
-e.g. "my_new_prototype" in `prototypes/models/my_new_prototype/...`
-"""
