--- corehq/apps/prototype/templates/prototype/webpack/bootstrap3_amd.html
@@ -1,21 +1,0 @@
-{% extends 'hqwebapp/bootstrap3/base_navigation.html' %}
-{% load i18n %}
-{% load hq_shared_tags %}
-
-{% js_entry_b3 "prototype/js/webpack/bootstrap3_amd" %}
-
-{% block content %}
-  {% initial_page_data 'test_initial' 'successfully retrieved initial page data' %}
-  <div class="container my-5">
-    <h2>
-      Testing an (HQ)AMD-formatted module loaded with Webpack
-    </h2>
-    <div class="alert alert-info">
-      FYI: This page uses Bootstrap 3.
-    </div>
-    <p>
-      Please open the js console for troubleshooting and see
-      "prototype/js/webpack/bootstrap3_amd.js" for additional information.
-    </p>
-  </div>
-{% endblock content %}
