--- corehq/apps/prototype/templates/prototype/webpack/knockout_pagination.html
@@ -1,49 +1,0 @@
-{% extends 'hqwebapp/bootstrap5/base_navigation.html' %}
-{% load i18n %}
-{% load hq_shared_tags %}
-
-{% js_entry "prototype/js/webpack/knockout_pagination" %}
-
-{% block content %}
-{% registerurl "prototype_example_paginated_data" %}
-  <h1>
-    Prototype Example (Webpack)
-  </h1>
-  <p class="lead">
-    This is an example endpoint for a rapid prototype working with tabular data and knockout.
-  </p>
-  <div class="alert alert-info">
-    This also serves as a more concrete example of migrating an AMD module to webpack and
-    updating the module  to use es6 module syntax.
-  </div>
-
-  <div id="prototype-example-knockout-pagination">
-    <table class="table table-striped table-hover">
-      <thead>
-        <tr>
-          <th>Name</th>
-          <th>Color</th>
-          <th>Big Cats</th>
-          <th>Date of Birth</th>
-          <th>Application</th>
-          <th>Opened On</th>
-        </tr>
-      </thead>
-      <tbody>
-        <!-- ko foreach: rows -->
-        <tr>
-          <!-- ko foreach: $data.columns -->
-          <td data-bind="text: $data"></td>
-          <!-- /ko -->
-        </tr>
-        <!-- /ko -->
-      </tbody>
-    </table>
-    <pagination data-apply-bindings="false"
-                params="goToPage: goToPage,
-                        slug: 'style-guide',
-                        perPage: itemsPerPage,
-                        onLoad: onPaginationLoad,
-                        totalItems: totalItems"></pagination>
-  </div>
-{% endblock content %}
