+++ corehq/apps/styleguide/examples/bootstrap5/htmx_complex_store.py
@@ -1,10 +1,10 @@
-from corehq.apps.prototype.models.cache_store import CacheStore
+from corehq.apps.styleguide.models import CacheStore
 
 
 class KeyValuePairStore(CacheStore):
     """
-    CacheStore is a helpful prototyping tool when you need to store
-    data on the server side for styleguide / demo HTMX views.
+    CacheStore is a helpful tool when you need to store data on the
+    server side for styleguide / demo HTMX views.
 
     Caution: Please don't use this for real features. It isn't battle-tested yet.
     """
